Backend/requirements.txt
Flask-Cors
flask
requests
python-dotenv
pandas
numpy